Add files via upload
added a quick option for functions.bulkprot, which was integrated into associated function calls/parameter definitions in main and cli.  quick option uses search instead of stream - quicker, but is limited in terms of the number of results returned.  Will need to implement pagination at some point, but currently main usecase is to scrape 21K entries via accession name, so no individual query should cause an issue with this